Add the `AlertManagerDefinition` Resource (#11)
Signed-off-by: Ilan <igofman99@gmail.com>

Issue #, if available: [#1323](https://github.com/aws-controllers-k8s/community/issues/1323)

Description of changes:

I think one change that requires extra attention is in `templates/hooks/alert_manager_definition/sdk_read_one_pre_set_output.go.tpl`. The behaviour of the API makes it tricky to deal async updates that result in failed states. I tried to explain everything in the comments, but if there are any other questions, please let me know. 

One of the special changes was adding a new `alertManagerConfig` field that is to be used which replaces the data field. From the API, the data field expects a base64 encoding of the rule groups YAML file. This is stored using a bye slice. However, providing a base64 encoding does not allow for readability or the ability to make changes quickly. Instead, we add a new configuration field that will be user facing. The user will submit the configuration as a multi-line string instead of using base64. The controller will handle the conversion to base64 before calling the APIs. An example of what the YAML file might look like:

```
apiVersion: prometheusservice.services.k8s.aws/v1alpha1
kind: AlertManagerDefinition
metadata:
  namespace: default
  name: ilan-alert
spec:
  workspaceID: ws-9db857ab-dc8b-46c3-ab83-4cdde703f867
  alertmanagerConfig: |
    alertmanager_config: |
      route:
         receiver: 'amp-topic'
      receivers:
        - name: 'amp-topic'
          sns_configs:
          - topic_arn: arn:aws:sns:us-west-2:742819168611:amp-topic
            sigv4:
              region: us-west-2
            attributes:
              key: key1
              value: value1
```

Instead of 
```
apiVersion: prometheusservice.services.k8s.aws/v1alpha1
kind: AlertManagerDefinition
metadata:
  namespace: default
  name: ilan-alert
spec:
  workspaceID: ws-0117031d-dd4e-4f20-8bcc-7e4bd5dc43f3
  data: YWxlcnRtYW5hZ2VyX2NvbmZpZzogfAogIHJvdXRlOgogICAgcmVjZWl2ZXI6ICdhbXAtdG9waWMnCiAgcmVjZWl2ZXJzOgogICAgLSBuYW1lOiAnYW1wLXRvcGljJwogICAgICBzbnNfY29uZmlnczoKICAgICAgLSB0b3BpY19hcm46IGFybjphd3M6c25zOnVzLXdlc3QtMjo3NDI4MTkxNjg2MTE6YW1wLXRvcGljCiAgICAgICAgc2lndjQ6CiAgICAgICAgICByZWdpb246IHVzLXdlc3QtMgogICAgICAgIGF0dHJpYnV0ZXM6CiAgICAgICAgICBrZXk6IGtleTEKICAgICAgICAgIHZhbHVlOiB2YWx1ZTE=
```
Note that right now the configuration is passed in as a delimited string (what the API expects). Inside the configuration, what the API expects, it is also composed of another delimited string. The next PR will remove the nested configuration values. Decided to keep things separate as it might get confusing putting all the changes to the custom field in the same PR. 

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.